consolidate variability analysis.
